Rename App.js to Calculate.js, create App.js and Quotes.js in src
