+++ src/App.js
@@ -1,13 +1,4 @@
-import './App.css';
-
-function App() {
-  return (
-    <div className="App">
-      <header className="App-header">
-        <h1>Hello World!</h1>
-      </header>
-    </div>
-  );
-}
-
-export default App;
+import React from 'react';
+import Nav from './Nav';
+import Calculator from './components/Calculator';
+import Quotes from './Quotes';
