+++ src/Nav.js
@@ -1,0 +1,14 @@
+import React from 'react';
+
+function Nav() {
+    return (
+        <nav>
+            <h1>Math Magicians</h1>
+            <ul>
+                <li>Home</li>
+                <li>Calculator</li>
+                <li>Quote</li>
+            </ul>
+        </nav>
+    )
+}
