+++ src/Quotes.js
@@ -1,0 +1,9 @@
+import React from'react';
+
+function Quotes() {
+    return (
+        <h3></h3>
+    )
+}
+
+export default Quotes
